Move convertlogic to textsService
