--- src/studio/src/designer/backend.Tests/Controllers/ConvertTextsControllerTests.cs
@@ -1,105 +1,0 @@
-using System;
-using System.IO;
-using System.Net.Http;
-using System.Threading.Tasks;
-using Altinn.Studio.Designer.Configuration;
-using Altinn.Studio.Designer.Controllers;
-using Altinn.Studio.Designer.Services.Interfaces;
-using Designer.Tests.Mocks;
-using Designer.Tests.Utils;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc.Testing;
-using Microsoft.AspNetCore.TestHost;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Moq;
-using Xunit;
-
-using IRepository = Altinn.Studio.Designer.Services.Interfaces.IRepository;
-
-namespace Designer.Tests.Controllers
-{
-    public class ConvertTextsControllerTests : IClassFixture<WebApplicationFactory<ConvertV1TextsToV2Controller>>
-    {
-        private readonly WebApplicationFactory<ConvertV1TextsToV2Controller> _factory;
-        private readonly string _versionPrefix = "designer/api/v1";
-
-        public ConvertTextsControllerTests(WebApplicationFactory<ConvertV1TextsToV2Controller> factory)
-        {
-            _factory = factory;
-            TestSetupUtils.SetupDirtyHackIfLinux();
-        }
-
-        [Fact]
-        public async Task ConvertTexts_204NoContent()
-        {
-            var targetRepository = Guid.NewGuid().ToString();
-            await TestDataHelper.CopyRepositoryForTest("ttd", "convert-texts", "testUser", targetRepository);
-            HttpClient client = GetTestClient();
-            string dataPathWithData = $"{_versionPrefix}/ttd/{targetRepository}/convertTexts";
-            HttpRequestMessage httpRequestMessage = new HttpRequestMessage(HttpMethod.Put, dataPathWithData);
-            await AuthenticationUtil.AddAuthenticateAndAuthAndXsrFCookieToRequest(client, httpRequestMessage);
-
-            HttpResponseMessage response = await client.SendAsync(httpRequestMessage);
-
-            try
-            {
-                Assert.Equal(StatusCodes.Status204NoContent, (int)response.StatusCode);
-            }
-            finally
-            {
-                System.Threading.Thread.Sleep(100);
-                TestDataHelper.DeleteAppRepository("ttd", targetRepository, "testUser");
-            }
-        }
-
-        private HttpClient GetTestClient()
-        {
-            string unitTestFolder = Path.GetDirectoryName(new Uri(typeof(DatamodelsControllerTests).Assembly.Location).LocalPath);
-            string projectDir = Directory.GetCurrentDirectory();
-            string configPath = Path.Combine(projectDir, "appsettings.json");
-
-            HttpClient client = _factory.WithWebHostBuilder(builder =>
-            {
-                builder.ConfigureAppConfiguration((context, conf) =>
-                {
-                    conf.AddJsonFile(configPath);
-                });
-
-                var configuration = new ConfigurationBuilder()
-                    .AddJsonFile(configPath)
-                    .Build();
-
-                configuration.GetSection("ServiceRepositorySettings:RepositoryLocation").Value = Path.Combine(unitTestFolder, "..", "..", "..", "_TestData", "Repositories");
-
-                IConfigurationSection serviceRepositorySettingSection = configuration.GetSection("ServiceRepositorySettings");
-
-                Mock<IRepository> repositoryMock = new Mock<IRepository>() { CallBase = true, };
-                repositoryMock
-                    .Setup(r => r.UpdateApplicationWithAppLogicModel(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>()))
-                    .Verifiable();
-
-                repositoryMock.
-                    Setup(r => r.ReadData(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>())).
-                    Returns<string, string, string>(async (org, repo, path) =>
-                    {
-                        string repopath = Path.Combine(unitTestFolder, "..", "..", "..", "_TestData", "Repositories", "testUser", org, repo, path);
-
-                        Stream fs = File.OpenRead(repopath);
-                        return await Task.FromResult(fs);
-                    });
-                repositoryMock.Setup(r => r.DeleteData(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>())).Verifiable();
-                repositoryMock.Setup(r => r.WriteData(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>(), It.IsAny<Stream>())).Verifiable();
-                repositoryMock.Setup(r => r.DeleteMetadataForAttachment(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>())).Returns(true);
-                builder.ConfigureTestServices(services =>
-                {
-                    services.Configure<ServiceRepositorySettings>(serviceRepositorySettingSection);
-                    services.AddSingleton<IGitea, IGiteaMock>();
-
-                    services.AddSingleton(repositoryMock.Object);
-                });
-            }).CreateClient(new WebApplicationFactoryClientOptions { AllowAutoRedirect = false });
-            return client;
-        }
-    }
-}
