+++ src/studio/src/designer/backend/Infrastructure/ServiceRegistration.cs
@@ -44,7 +44,6 @@ namespace Altinn.Studio.Designer.Infrastructure
             services.AddTransient<ISigningCredentialsResolver, SigningCredentialsResolver>();
             services.AddTransient<ILanguagesService, LanguagesService>();
             services.AddTransient<ITextsService, TextsService>();
-            services.AddTransient<IConvertTextsService, ConvertTextsService>();
 
             return services;
         }
